[PR] Update Readable definition so that _read always takes a size
Summary:
According to the documentation, `_read` takes a non-optional size as argument. Marking it non-optional still allows to extend the class and ignore it, so this should be backward compatible. However this change makes it possible to handle this argument without checking the presence first.

Documentation links:
* https://nodejs.org/docs/latest-v12.x/api/stream.html#stream_readable_read_size_1
* https://nodejs.org/docs/latest-v12.x/api/stream.html#stream_new_stream_readable_options

I looked as far as node v4 which has the same documentation.

Pull Request resolved: https://github.com/facebook/flow/pull/8385

Reviewed By: panagosg7

Differential Revision: D22597255

Pulled By: mroch

fbshipit-source-id: 513e6a9aee01af49e2831f1b52d019444e7f843c