Check arch before fetching sig_file_sig
Summary:
There's a bit of code in Dep_service that calls get_sig_file_sig, then checks
the returned value for None. The idea here is that, if the return value is None,
then there is not a signature.

This invariant makes sense as long as there is only Classic and TypesFirst
architectures, but I am going to add a third architecture temporarily as we
migrate to TypesFirst 2.0.

Instead, this diff replaces the option check with a check on Options.arch. If
the arch is TypesFirst, then we can do an unsafe lookup for the sig_file_sig.

Reviewed By: mroch

Differential Revision: D22577528

fbshipit-source-id: 4d3637b186ca5d351eec6a1664ca429c04ffa266