Fix: update ServerEnv errors when initializing from saved state
Summary:
Fixes the bug described in D22847292.

Before, when initializing from saved state, the server environment
was populated by
1. restoring information from saved state
2. calling `parse_and_update_dependency_info` on the files that changed in
the meantime (these are reported in the "saved_state_file_changes" file).

Errors found in (2) were not recorded in the env. The `intermediate_values` from that
step were just dropped. (This did not happen in the lazy-mode initialization codepath,
where `env` is updated with the new errors.

This diff changes this by updating `env.SeverEnv.errors` with the local errors found
during (2).

Reviewed By: samwgoldman

Differential Revision: D22873641

fbshipit-source-id: 96dd39c0cda91a9bf37a8d472624f6138a6433c4