Avoid Hashtbl.find_all API
Summary:
The OCaml Hashtbl module is somewhat unusual -- `add k v1` followed by `add k
v2` will leave both `v1` and `v2` in the bucket for `k`, and `find_all` will
return a list containing both `v1` and `v2` by scanning the entire bucket.

While this is a possible way to use the Hashtbl module, it's not a very *good*
way to use it. This diff accumulates a map from modules to list of dependent
files instead. Furthermore, `find_all` is not tail-recursive, so if there are
many values in a given bucket, this call can lead to stack overflows.

Reviewed By: mroch

Differential Revision: D22861902

fbshipit-source-id: 7ea2307252bad055cd7ab8352537a7d0a498c9a3