Support running tests in new signatures mode
Summary:
This diff adds a new flag to the runtests.sh test runner, which toggles tests to
run in new signatures mode (passing the --new-signatures cmdline flag).

This is similar to the flag that toggles tests to run under saved state.

Many tests are currently failing, often for minor differences in reason
descriptions or locations. Tests which are failing are configured to be skipped
via the TODO_new_signatures .testconfig setting.

The existing classic_only setting also disables these tests, since any test that
can not run under types-first will also not run with new signatures.

Reviewed By: mroch

Differential Revision: D22938923

fbshipit-source-id: cec7093404b0d539c8c1a72ef8af3f645f8eef99