Add merge implementation for new signatures
Summary:
This diff changes the behavior of the parsing and merge phases of type checking
when the experimental.new_signatures option is enabled.

When this option is enabled, parsing will create type sig data and save that to
shared memory. Merge will read those type sigs from shared mem and use them
to build signature contexts.

Reviewed By: panagosg7

Differential Revision: D22875638

fbshipit-source-id: 90ed69fb41d0453310655a20dbd237d9c9ac68c7