Add tests for method locations
Summary:
The location associated with the FunT for a method is a bit inconsistent. These
tests simply document the existing behavior; later diffs will fix.

Note that I was careful to test both inter- and intra-file errors, since the
types-first architecture makes it possible for those to differ. For example,
observe that we say "function" vs. "function type" because of the way that
types-first encodes signature ASTs.

Reviewed By: panagosg7

Differential Revision: D22971883

fbshipit-source-id: 3b0afdea344ced20d3bbfb81fc612e637dbc33dc