Refactor parsing service to make room for new signatures mode
Summary:
Currently both classic and types-first modes share code in the parsing code
path. Specifically, both modes perform signature verification, generate sig
ASTs, and include signature verification errors in the tolerable errors set.

This diff lifts the branching condition to the top-level of this code, making
room for a third branch for new signatures, which does not share the same code.
In practice, this does not lead to much additional code duplication, as the
classic mode branch can be simplified.

In fact, this change removes a potential inefficiency in classic mode, as we no
longer generate sig ASTs in that mode. Note that we still perform signature
verification, as signature verification errors will be emitted if the
well-formed exports option is enabled (filtering is done later).

Reviewed By: panagosg7

Differential Revision: D22875637

fbshipit-source-id: 94f8ab1fa34ecdec82230ac32c2d6ab8c59cf898