Make function locations consistent between classic and types-first
Summary:
Due to the way that types-first creates signature ASTs, it's actually really
tricky to coax it into creating types that exactly mimic the ones which would be
created from the full AST.

For example, function declarations are converted to declared functions, which go
through different code paths and end up with different reasons. This is why the
reason descriptions are "function" in classic and "function type" under
types-first.

This diff makes sure to use the "right" location in key places in the signature
AST, so that the types created during merge will point to the correct locations.
Generally this means using the "sig_loc" field of the function AST node.

Reviewed By: panagosg7

Differential Revision: D22971885

fbshipit-source-id: de9cb55cea0ead4b903036313d3c268b8b5a5f6a