Add --new-signatures command line option
Summary:
This comand line option makes it easier to test the experimental new signatures
mode. Before this, it was necessary to modify the .flowconfig file to turn on
the mode.

In particular, this will make it easier to run automated tests. Instead of
modifying the .flowconfig for each file, we can tell the test runner to pass
this flag instead.

The new signatures mode will hopefully replace sig ASTs soon, making this option
useless. I did not include docs for the option, which prevents the option from
being printed along with --help / usage.

Reviewed By: mvitousek

Differential Revision: D22927954

fbshipit-source-id: b6fba5f7674fe1f22f458f41926196e60ad68886