Fix location of type params
Summary:
This diff matches the behavior in Type_annotation, which uses the location of
the type parameter name. Before we were using the location containing the
variance sigil, name, bound, and default.

I also changed the MixedT that we create for unbound generics to match the type
created in Type_annotation exactly. Before this change, the MixedT.at call would
use create a type with a different reason description.

Reviewed By: mroch

Differential Revision: D22939345

fbshipit-source-id: b66ce84d9ca195f42c322e1a8575cf59d4cc35bb