Make the location of method types consistent between introduction forms
Summary:
We create FunT types to represent methods in classes, declared classes, objects,
object types, and interfaces. Currently, the location associated with those
types is inconsistent, pointing to either `m() {}` or `() {}` (excluding the
identifier).

This diff changes all forms to include the location of the identifier -- i.e.,
use the location of the entire property.

Per discussion with mroch, it's problematic to point to the body of a function
definition, since it usually spans multiple lines. The location used after this
change is still problematic.

For future work, we should look into using just the identifier name where
possible. This would avoid the case where an error spans multiple lines.

Reviewed By: mroch

Differential Revision: D22971884

fbshipit-source-id: cb719aedca65bc815a36aa9462b66ac6c9d85085