Move ES6 import/export error generation alongside other post-merge errors
Summary:
Previously, ES6 import/export errors were being generated in `Merge_service.merge_context_generic`, which is a different location than other post-merge errors are generated. This is causing issues with the IDE integration, since these errors are being added after the merge job sends its errors to the persistent connection. This is resulting in the ES6 import/export errors not being shown in the IDE, even though other errors are shown in the IDE.

Let's move the ES6 import/export errors to be detected alongside the other post-merge errors (in `Merge_js.merge_component`), so that they are displayed correctly in the IDE.

Reviewed By: panagosg7

Differential Revision: D22931907

fbshipit-source-id: 0e68e6d4f8e2ca4a80608ce19b79bb68c1802f73