exit workers gracefully when server dies
Summary:
if the server dies, each worker will get hit by an `EPIPE` when it tries to write its response to the broken connection.

the server should not go away unexpectedly, but the monitor already watches for that, so it's redundant for each worker to also spew errors.

the workers already caught `End_of_file` (trying to read the request from a dead server) and exited gracefully. this diff does the same for writing the response.

Reviewed By: nmote

Differential Revision: D23079294

fbshipit-source-id: 70cfc0cd43b2d5788949c098a450faceb02b54da