Change function reasons to match existing behavior
Summary:
This change ensures that types-first 2.0 will produce types that have the same
reason as the types produced by types-first 1.0. Before these changes, the
locations and descriptions of those reasons could differ.

For function declarations and non-arrow function expressions, we will use the
"sig loc," which includes the name, parameters, and return type but not the
body. For arrow functions we use the entire span of the expression. For declared
functions we use a location including the parameter list, return type, and body.

Separately, it would be nice to make this different syntactic forms consistent
between themselves. For now, I'm satisfied to make the handling of these forms
consistent between the various modes -- classic, types-first, and new sigs.

Reviewed By: panagosg7

Differential Revision: D22971882

fbshipit-source-id: 2e24870d9d28d66962e8d5208dd7449d8f404120