Avoid stack overflow on very large types
Summary:
List.map in the OCaml stdlib is not tail recursive, but in extreme scenarios
some lists might be large enough to cause stack overflows. This diff ensures
that those cases will use the tail-recursive map.

This is a conservative change. I could have tried to change everything to use
Base.List instead, but holding off on that for now. Would want to do some actual
benchmarking on a bigger change.

It would also be interesting to see how this compares to TRMC[0], which might
find its way into a future version of the stdlib.

0. https://github.com/ocaml/ocaml/pull/9760

Reviewed By: panagosg7

Differential Revision: D23152265

fbshipit-source-id: 14d2a575d534d9c8cab73b89d0e6f1a52598f046