run literal conditional checks post inference
Summary:
This change is related to a check that Flow performs along with conditional
refinement. This check ensures that the "sentinel" value inhabits the "refined"
value.

There are two common forms of this check
* primitive value refinement (`x === "value"`)
* sentinel property check (`obj.property === "value"`). This was
introduced in D5971535 (https://github.com/facebook/flow/commit/e4451324153aa7b73d235c48dff47a4aa3b10fd9), and it is the reason for introducing the
`MatchingPropT` type.

The problem with the previous approach is that this check was done during
constraint propagation, and so it could affect the result of inference, by flowing
the "sentinel" type to the type of the refined value.

This can potentially add the sentinel type as a lower bound to type variables,
and cause it to show up downstream, far from the conditional, in the form of
spurious errors. For example,
```
const {key} = {key: 'a'};
const x = key === 'b' ? null : null;
(key: 'a');
```
Before, the comparison of `key` with `'b'` caused the check in the third line to fail.

In this diff, we move these checks to be performed after inference has concluded,
as a "post-merge" pass. There's two changes that achieve this:
1/ We keep track of the tests that need to be performed in a list in Context.t.
2/ After inference is done, we apply the "merge-tvar" function on the involved
types that we have gathered and then fire the constraints.

Essentially, this allows us to remove the effect of "widening" tvars past the point
that we care about in constraint propagation, but at the same time reuse existing
mechanisms to do the subtyping checks.

I've also removed this check from heap refinements, in favor of using the attendant
SentinelProp mechanism. This reduces some earlier duplicate errors.

Reviewed By: samwgoldman

Differential Revision: D22834117

fbshipit-source-id: 7dae8122d83de5df935f4f032bd700f69e002238