[refinements] add tests on type expansion due to inclusion checks
Summary:
A check that Flow performs along with conditional refinement, is that the
"sentinel" value inhabits the "refined" value. For example, in
```
const x: { type: 'A', ... } | { type: 'B', ... } = ...;
if (x.type === 'C') {
  ...
}
```
we error at the conditional check. This check is not necessary for soundness,
but it helps prevent some common bugs.

The caveat with this approach is that this check, when done during inference,
might affect the result of inference, by flowing the "sentinel" type to the type
of the refined object. This can potentially add the sentinel type as a lower bound
to type variables, and cause it to show up downstream, far from the conditional, in
the form of spurious errors.

This diff adds tests that exhibit this behavior.

To avoid erroring in these case, later on, we move this check to a post-inference
pass.

Reviewed By: samwgoldman

Differential Revision: D23025094

fbshipit-source-id: 9e90f04e8bd10e3fe8c6779255a47de6b7189224