do not apply sentinel-prop refinement on Switch case expression
Summary:
To compute the refinement happening in
```
switch (exp) {
  case val: ...
}
```
we create "fake" predicate expressions of the form `exp === val`. The work as
expected when `val` is not a member expression, but may cause an unwanted
sentinel prop refinement if `val` is of the form `o.p` (the sentinel refinement would
be on `o` if `exp` is not a member expression itself.

This diff bypasses this refinement by adding a check in `predicates_of_condition`.

Reviewed By: samwgoldman

Differential Revision: D23155675

fbshipit-source-id: 8129aa245e1a20b4f119b06be42c9ea72a3f3fa2