enable EventLogger in workers
Summary:
`EventLogger` needs to be initialized in each process before it will log anything. we initialize it directly in the monitor when we start the monitor with `flow server`, and also in the server process that the monitor spawns. but we don't currently initialize it in each worker. this causes us not to log `WORKER_EXCEPTION` events when a worker hits an uncaught exception :facepalm:

here, I'm following Hack's approach of having each worker call `EventLogger.init` with the same `init_id` as the server, plus a unique ID per worker (see `FlowEventLogger.init_worker`).

`EventLogger.init` spawns a logger process, so we end up with one per worker (:-O) but I don't see much downside and it's what Hack does, so ¯\_(ツ)_/¯

Reviewed By: panagosg7

Differential Revision: D23068870

fbshipit-source-id: 508b1cbde199d513cd9da26b539637860677570a