enable EventLogger in workers
Summary:
`EventLogger` needs to be initialized in each process before it will log anything. we initialize it directly in the monitor when we start the monitor with `flow server`, and also in the server process that the monitor spawns. but we don't currently initialize it in each worker. this causes us not to log `WORKER_EXCEPTION` events when a worker hits an uncaught exception :facepalm:

here, I'm following Hack's approach of having each worker call `EventLogger.init` with the same `init_id` as the server, plus a unique ID per worker (see `FlowEventLogger.init_worker`).

`EventLogger.init` spawns a logger process, so we end up with one per worker (:-O) but I don't see much downside and it's what Hack does, so ¯\_(ツ)_/¯

Reviewed By: panagosg7

Differential Revision: D23068870 (https://github.com/facebook/flow/commit/58261b16d6ea87a19702b6c2c06d4216caba2457)

fbshipit-source-id: 97b05f57bc6c817655a2615c6f81a4fcdf110b1f