JSDoc: parse advanced @param tags
Summary:
According to the [JSDoc documentation](https://jsdoc.app/tags-param.html), the parameter name in an `param` tag can be things other than simply an identifier:
https://jsdoc.app/tags-param.html
- `[foo]` means that the `foo` parameter is optional
- `[foo=bar]` means that the `foo` parameter is optional and has default value `bar`
- `foo.bar` documents the `bar` property of the parameter `foo`
- `foo[]` documents the array elements of the parameter `foo`
A parameter can be referred to multiple times through different "destructors" to document different parts of it, and various parts of a parameter can be documented as optional. e.g.
```
/**
 * param foo - description of foo overall
 * param foo.bar - description of the bar property of foo
 * param [foo.baz] - description of the baz property of foo, which is optional
```
This diff updates the parser and JSDoc data structure to support this. The info of each parameter is stored as a map from a `path` (representing a possibly empty sequence of destructors) to an `info` record which stores the parameter's description (if present) and whether the parameter is optional (and if it is optional, the default value).

Reviewed By: mroch

Differential Revision: D22782512

fbshipit-source-id: 61b89a163410bd37f771a03fd0609da3849374da