JSDoc: preserve order of @param annotations
Summary: switch to using assoc lists instead of maps to store info about param tags, so that we know what order they were written in. This will be useful for matching param annotations to the function's parameters, and also for rendering the info to the user in the same order it was written in.

Reviewed By: mroch

Differential Revision: D22957251

fbshipit-source-id: 8ca8c82c2a51d7cf5050977b5950dbfb6ef6ed5c