Fix crash related to saved state and deleted files
Summary:
If a file is deleted at the wrong moment, after a recheck has started, it can lead to a crash. Here's how this happens:

1. Flow starts from saved state, parsing nothing on startup.
2. A recheck starts, and Flow determines that it needs to parse some previously unparsed dependencies.
3. One of those dependencies has been removed, but Flow hasn't found out yet (this can happen for a variety of reasons, but it mostly boils down to race conditions).
4. `ensure_parsed` is called with the removed file as one of the files to parse.
5. Parsing of that file fails because it is no longer on disk, and therefore the parsing artifacts do not make it into the parsing heaps.
6. When the `merge` and/or `check` phases try to pull these artifacts out, we get a crash.

The issue here is that when `Parsing_service_js` is asked to parse a nonexistent file, it logs an error, and adds the file to the list of skipped files. Typically, skipped files are files that are **intentionally** skipped, and not skipped because of some issue. `ensure_parsed` ignores the list of skipped files, because they are usually innocuous. Instead, it should treat them more like the set of files that have hash mismatches, indicating that they have changed and the recheck needs to start over.

I added a category in the parse results data structure for files that were skipped because they were not found. Then I updated all the callsites (except for `ensure_parsed`) to treat those files as if they were in the `parse_skips` field, so as not to change unrelated behavior. In `ensure_parsed`, I changed it to treat this new category the same way as `parse_hash_mismatch_skips` is treated, thereby triggering a fresh recheck when this happens, and avoiding the crash.

Reviewed By: mroch

Differential Revision: D23110950

fbshipit-source-id: 1a47ec9767ee7fca1e5988d5f2d9243933e5fe32