do less work when walking directories
Summary:
a directory that is not a symlink is guaranteed to be within the search directory, so the `prefix_checkers` check on non-symlink dirs will always return false.

a directory that _is_ a symlink is added to a list to visit later, so we don't need to `readdir` it (and throw away the results).

Reviewed By: samwgoldman

Differential Revision: D23196459

fbshipit-source-id: f899bc02d6fe4041bcdce62abcb235f895f075c6