Fix Lwt.Canceled crash
Summary:
`ensure_parsed` is used after loading from saved state, since with saved state we do not parse every file. Operations that require certain files to be parsed must call it. Previously, if `ensure_parsed` encountered an unexpected file change (either a file that does not match the hash that was stored in saved state, or a file that is no longer on disk), it would enqueue a new recheck and raise `Lwt.Canceled`.

This worked fine when we were inside of a recheck, since the ongoing recheck would be canceled (`Lwt.Canceled` exceptions are handled up the stack), and a new one that includes the surprise changes would be started. Unfortunately, we call `Recheck.parse_and_update_dependency_info` on saved state init, in order to process the dependency graph changes from files that have changed since the saved state was generated. As part of updating the dependency info, it needs to call `ensure_parsed`. If it encountered any unexpected file changes, it would enqueue a new recheck, and raise `Lwt.Canceled`.

The problem with this is that during init, Flow isn't in the context of a recheck, so there was nothing to catch the `Lwt.Canceled` exception, and Flow would crash. Worse, even if we caught it, Flow would end up in a weird state where it hadn't actually updated the dependency info for all of changed files, and then it would immediately after init do a recheck of the files that `ensure_parsed` noticed were changed.

So, I introduced a custom exception that includes the list of changed files, so that if `ensure_parsed` encounters an unexpected change, we can incorporate it into the set of changed files that we learned about when loading the saved state, and just finish the lazy init normally.

Reviewed By: mroch

Differential Revision: D23222902

fbshipit-source-id: aa326490e6ba7518115ede68b7a9b77948113d22