Raise a custom exception from `ensure_parsed` when files have changed
Summary:
Previously we would just trigger a new recheck and then raise `Lwt.Canceled`, which meant that callers had no details about the cancellation. This works fine when we are calling `ensure_parsed` from inside of a recheck, but it does not work so well when we call it during init, where some details about the failure would be useful (see the next diff).

I'm not a huge fan of using exceptions in this manner, but we were already using `Lwt.Canceled` for this purpose, so I feel that simply changing it to a different exception that carries more information is not much worse. I thought about running another monad with this cancellation information all the way through `recheck_parse_and_update_dependency_info` instead, but we already have `Lwt.t` in the mix and I've been down the double-monad road before. It's not a fun place to be.

Reviewed By: mroch

Differential Revision: D23222903

fbshipit-source-id: b7e0a5785eb56681f6cc0c7db69f9e67d2779049