packed locs
Summary:
In the new heap, we store an array of locations for each checked file. In the
signatures, we store offsets into this array.

Before this diff, we were storing just a `Loc.t array`, serialized using the
Marshal module.

The biggest factor was that each Loc.t stores the source file's filename. A
given array will only serialize the filename once, due to Marshal's sharing
support, but it's unnecessary because we already store the filename in the heap.

Just removing filenames saves the most space in practice, but I got a bit
obsessed and decided to write a custom encoding to crush things down even more.
See the docblock comment in packed_locs.ml for more details.

Reviewed By: mroch

Differential Revision: D23173881

fbshipit-source-id: 5fc22c02f4b2a48ab113b806d0f17778f30972b6