refactor: functorize codemod_runner
Summary:
Using modules to describe the various codemod visitors is more
natural than records, since modules let you define types specific to
each transformation, and are more flexible and composable.

The code in codemod_runner already used the `RepeatRunner` functor
to convert single codemod steps to repeatedly applied codemods.

This change replaces
```
type 'a visitor =
  | Typed_visitor of ('a, Codemod_context.Typed.t) abstract_visitor
  | UntypedFlowInitRunner_visitor of 'a untyped_flow_init_runner_visitor
  | Untyped_visitor of ('a, Codemod_context.Untyped.t) abstract_visitor
```
with three new module signatures
* `SIMPLE_TYPED_RUNNER_CONFIG`
* `UNTYPED_FLOW_INIT_RUNNER_CONFIG`
* `UNTYPED_RUNNER_CONFIG`

in codemod_runner.ml. The rest of the diff propagates the change the various
files that call it.

The main motivation for this change is a new configuration that I'm planning
to add,  that runs in two phases: a pre-pass that gathers checking information
from downstream uses, and a codemod pass that uses this information to
apply the codemod. With our previous setup it would have been awkward
making sure that the type collected during the pre-pass matches the one
read from codemodding pass.

Reviewed By: Hans-Halverson

Differential Revision: D23168884

fbshipit-source-id: ed2fe85eac60612f238a916e481aa98a02140405