Add typed pre-pass runner in codemods
Summary:
This diff adds a new mode in the codemod runner. This mode includes two typed
passes:
1. The first pass is intended to gather information from downstream uses of the
input "root" set. This means that we have to check not only the "root" set, like we
do in SimpleTypedRunner, but also the set of dependents of the "root" set.
Thanks to this change in the spec, there is little code to be reused between the
two modes. Notably, `merge_targets` needs to be changed to reflect this: the
set of files to be merged/checked resembles the one we compute on watchman-
lazy mode, for example.

2. The second pass resembles the one in SimpleTypedRunner. We only check
the "root" set. At that point it should have been merged already.

This mode does not determine how information will be passed from the first
pass to the second. It does provide two callbacks, however:

* `prepass_init : unit -> prepass_state`
* `val store_precheck_result : prepass_result unit_result FilenameMap.t -> unit`

The expectation is that the client analysis will store information in main memory
right after the first pass.

See D23168956 for an example of how this mode can be used.

Reviewed By: Hans-Halverson

Differential Revision: D23232090

fbshipit-source-id: 28fa6ed5ea3a8389df362bb284286700a26507f6