fix crash with `flow focus-check` when a focused file has a syntax error
Summary: `flow focus-check` will crash when a focused file has a parse error. This doesn't happen with regular unfocused `check` because in that case the focused set is all parsed files; the problem occurs because with `focus-check`, we can get an unparsed file in the checked set. This diff fixes that behavior by removing unparsed files from the focused set.

Reviewed By: mroch

Differential Revision: D23285418

fbshipit-source-id: 94a0352122dcacb93407ea8beec8c154032c3f89