allow Find_documentation to take the current command's AST
Summary: Make find_documentation take an optional `current_ast` argument which will be the typed AST of the file the user has open. This partially fixes a bug where if there are parse errors in a file we won't find JSDocs in that file. This diff will allow us to find JSDocs in the current file even if there are parse errors in the current file. We still won't be able to find JSDocs in other files with parse errors, but this should be ok since in most cases when there are parse errors in a file it's because the file is currently being worked on.

Reviewed By: mroch

Differential Revision: D23330116 (https://github.com/facebook/flow/commit/bbce60e5d74950ab0f2f2e5fbd1fdccb8014006a)

fbshipit-source-id: e24b3ca41177b0f7599914e315beda1d62917fd1