Remove coverage map from saved state
Summary:
This is now unused -- see previous diff.

This reduces the size of the saved state, but only when the server generating the saved state has focused files (in lazy mode), or if it is a non-lazy server. This is because the coverage map is only populated during typechecking.

It's also worth noting that the paths in this map were never normalized and denormalized like everything else in saved state, so the restored coverage maps would use the original paths of the files as they were on the machine that generated the saved state. This means that if we had been relying on them (we weren't), the things relying on them would have been broken anyway.

Reviewed By: mroch

Differential Revision: D23402055

fbshipit-source-id: 8ef1c14aa02064f24c75a799f9b28106aebf3180