Combine use_ops correctly
Summary:
This diff ensures that we consider the same use_ops in `flow_use_op` and produce the same final result regardless of the order in which we see the constraints.

To do this, we needed to start tracking use_ops on the upper/lower tvars in constraints, as well as call `pick_use_op` (the mechanism `flow_use_op` uses to pick a use_op) in a few more places.

As always, changing the use_ops here causes some error messages to get worse and others to get better. There is no way to make the errors "good" in all cases. Consider these two examples:
1. [printBackwards](https://flow.org/try/#0PTACDMBsHsHcCgAmBTAxpAhgJ2QAnAK4B2qALgJbRG4AOW5RpAFNgOYBcuAzqfUawEpOAN2jlEAbnjTCJClVp9SAIQyoA1rGyIuTaANwBveLkUNm++AF8kaTDlzDsuAB6dDuAJ6ciBALYARshYuFZSdOaqGlpYOkwuAHSeAhJAA)
2. [printNegation](https://flow.org/try/#0PTACDMBsHsHcCgAmBTAxpAhgJ2QAnAK4B2qALgJbRG4AOW5RpAyqfUQOYAU27AXLgGdWDdgEp+AN2jlEAbngLCJClVptSAOWTsMKop2ijcAb3i41DZsI4HR8AL5I0mHLgnZcAD37 (https://github.com/facebook/flow/commit/78b31ee637559cd89b3dda0f5cce4e58ebde11bc)HcAT34iAgBbACNkLFx7eTpLLR09Tk8AOj9RWSA)

In 1., the error is positioned poorly because we are calling `printBackwards` with a `number`. We should be positioned at that call!

In 2., the error is positioned well. The issue isn't that we are calling `printNegation` with a number, it's that the implementation of `printNegation` is wrong.

These two examples are identical in every way except for the names of the functions, which makes it clear that positioning an error message in the best place is not something flow can do with the information it has.

The benefit this diff brings is consistency-- when encountering constraints in different orders we will always produce the same use_op, which is necessary for the new inference project.

Reviewed By: panagosg7

Differential Revision: D23009760

fbshipit-source-id: 0bfdf1bab58714cf9f762bd8923c82ea61756199