[Refactor] refactor type definitions to restrict GetPropT and related use types to have tvars in the tout position
Summary:
We often assume that `tout`s are going to refer specifically to `OpenT`s, but have no actual enforcement of this in Flow's internal type system.

For the openness project we are going to have to determine precisely where violations of this assumption occur, so can here put OCaml's type system to work to refactor use types to enforce that their `tout`s are indeed type variables by changing the output type to `Type.tvar` instead of `Type.T`.

Reviewed By: jbrown215

Differential Revision: D23430684

fbshipit-source-id: a27c2eef04b27160bd97c63bbb105e8cb035444e