Migrate ensure_parsed_no_saved_state test to types-first
Summary:
This test initially failed when I enabled types-first on it because the last call to `flow status` unexpectedly returned "No errors"

I didn't investigate really carefully to find out what was going on here, but I suspect that it's something along these lines:
* `C.js` is never focused. It's only pulled in as a dependent of `A.js`
* Flow gets a `force-recheck` request for `A.js` just before this last `force-recheck` request, but since `A.js` hasn't actually changed it short-circuits early and doesn't check downstream.
* Flow gets a `force-recheck` request for `C.js`, but since it's not focused, and no upstream files changed, it skips the `check` phase.

Anyway, regardless of what's going on, I think this is a reasonable change because `force-recheck --focus` emulates what happens in lazy mode (both fs and watchman) when the filesystem reports that a file has changed. That file then becomes focused. In my opinion, this test as it previously existed was well into the territory of undefined behavior.

Reviewed By: samwgoldman

Differential Revision: D23607583

fbshipit-source-id: 64a346603842e7e680cde80e9edf784ef02c5faa