When switching on a union, if discriminant is empty at end, don't add implicit void return
Summary:
In D21814721 (https://github.com/facebook/flow/commit/d7711aa9e4a1cd0b5e55f6abbcb1f07797e142ab) I added support for basing whether a function implicitly returns void on whether an exhaustive check on a Flow Enum was successful.

In this diff, I add support for unions. More specifically, when the switch discriminant (what we're switching on) is refineable (an identifier or property), then each `case` refines the value of the discriminant. If we return from each case, and the switch does not have a `default`, we can check whether the type of the discriminant after the switch is empty. If it is, then we don't add the implicit void to the function return. This works because `EmptyT` < `FunImplicitVoidReturnT` will succeed without the additional flows `FunImplicitVoidReturnT` triggers (flowing void to the return type).

Currently, people have to add a `default` case and throw from it, or use other similar patterns.

E.g. the following is now allowed:
```
function f(x: 'a' | 'b'): number {
  switch (x) {
    case 'a':
      return 1;
    case 'b':
      return 2;
  }
}
```

Reviewed By: samwgoldman

Differential Revision: D23202941

fbshipit-source-id: 3494c66d5898630f93c8ae3ae7827ddf2aa1ec5c