add config option for commit to compute watchman mergebase against
Summary:
when using watchman lazy mode, flow checks all files that have changed since the "mergebase", the point at which your stack of local commits diverged.

we currently hardcode `master` as the symbolic commit to compute the mergebase against (so, all changes between `master` and `.`).

this diff adds a `file_watcher.watchman.mergebase_with` flowconfig option and `--file-watcher-mergebase-with` CLI option to choose a different tracking bookmark (say, `main`), or even a specific commit if you only want to check changes on part of a stack.

Reviewed By: jbrown215

Differential Revision: D23636608

fbshipit-source-id: a41d9bf23a095e25c6f0333c63616bb9e01658f3