[enums] Fix issue with exhaustive checks and implicit void return
Summary:
Whether a function implictly returns void depends on whether its switch on Flow Enums are exhaustive. We had an issue where in the following case we were reporting (incorrectly) that the function implicitly returned void:
```
function f(x: ?E): string {
  if (x == null) {
    return 'c';
  }
  switch (x) {
    case E.A:
      return 'a';
    case E.B:
      return 'b';
  }
}
```
This is because `x` is an `OpenT`, after the refinement with `x == null`, its lower bounds are `EmptyT` and `EnumT`.

When we have `EmptyT, EnumExhaustiveCheckT`, we don't automatically succeed without further work. This is because we use `EnumExhaustiveCheckT` to resolve the switch cases and the discriminant (and finally do the exhaustive check), and we want to keep the following error which is relevant when resolving the switch cases:

```
master/fbcode/flow/tests/enums/exhaustive-check.js208 function f11(x: E, X: empty) {
209   switch (x) { // Error
210     case X.A: break;
211   }
212 }
```

However, we were not handling empty in the not-resovling-switch-cases case.

By adding logic to explicitly handle this case, we resolve our original issue, and we properly flag the function as not having an implicit void return.

This change does not affect whether we consider an empty discriminant to be a succesful exhaustive check, as it already is due to D23202941 (https://github.com/facebook/flow/commit/96868b5372f47771876db63741351f89ca02562e).

Reviewed By: panagosg7

Differential Revision: D23665981

fbshipit-source-id: d900c6a05ccf0fb2346bd12635fa4dbddb2c0031