Enable use of TypeScript's lib defs source in our lib defs
Summary:
Flow recently added the ability to display JS doc comments.

We would like to add such comments to our library definitions. TypeScript's library definitions already have such comments written, so we would like to use theirs (with modifications if required). TypeScript is under the Apache 2 license.

I have talked to the lawyers, which said this was fine (as MIT and Apache are compatible licenses), as long as we 1) include a copy of the Apache license 2) include Microsoft's copyright notice at the top of our files, along with a "modification notice". I have done those things, and as an example have added TypeScript's jsdoc comment for `isNaN`.

Reviewed By: indirabhatt

Differential Revision: D23674494

fbshipit-source-id: 9b62da2e062b4d6af789fdd98403a02b44ba7dcc