add config option for commit to compute watchman mergebase against
Summary:
when using watchman lazy mode, flow checks all files that have changed since the "mergebase", the point at which your stack of local commits diverged.

we currently hardcode `master` as the symbolic commit to compute the mergebase against (so, all changes between `master` and `.`).

this diff adds a `file_watcher.watchman.mergebase_with` flowconfig option and `--file-watcher-mergebase-with` CLI option to choose a different tracking bookmark (say, `main`), or even a specific commit if you only want to check changes on part of a stack.

Reviewed By: jbrown215

Differential Revision: D23636608 (https://github.com/facebook/flow/commit/bff288a0bb05d6afe7ecea614688440544db203d)

fbshipit-source-id: 481272d56d55e0f98b288bba5cf8350d36b56084