[new-signatures] Use Signature_error.t when possible for better error messages
Summary:
This diff makes the signature verification errors generated by TF2.0 much closer to those produced by TF1.0. This is accomplished by the `type_sig_parser` directly generating the same error values that generated by TF1.0, using the `SigError` error value introduced in the previous diff. For the most part, this is accomplished by just replacing some of the `Type_sig.errno` errors that existed before with equivalent `Signature_error.t` errors. In some cases this requires more information--mostly locs--to be threaded through the parser.

From my (not exhaustive) testing, after this diff, there's three ways in which the signature errors from 1.0 are different from 2.0, which are documented in `tests/signature_verification_error`, which I've added a bunch of tests to:

* Only 2.0 errors: this is the largest group, and consists of cases where Types First 1.0 doesn't report a signature verification error (though it usually does show some other error) but TF2.0 does. This occurs for, example, with incorrect arity, builtin type operators. `export var b: $TEMPORARY$number<1,1> = 42;` is a verification error in Types First 2.0, but not in 1.0 (in both cases it's also a `nonpolymorphic-type-app` error)

* Only 1.0 errors: this comes up in a few cases like computed and private class properties: `export class C { #m = 42 }` will raise a signature verification error for `#m` in TF1.0, but not in 2.0. It seems like the 2.0 behavior is sound; external modules can't view these properties so why should they be in the signature? But it is a difference in behavior.

* Different error behavior: the only place I know of that still has different error messages, but both versions error, is for exported objects and array literals. `module.export = { a, b }` will give two errors pointing at `a` and `b` in TF1.0, but in 2.0 will have a single error pointing at the object `{ a, b }` as a whole. Fixing this is the next bit of work here, and looks a little more complicated than the rest, so I'm planning on working on it in a separate diff.

This diff also removes the `TODO_new_signatures` from lots of test cases which pass now.

Finally, one note is that while this diff duplicates the TF1.0 behavior as well as possible, that behavior does leave a lot to be desired: for example, the error for `export function f(x): void { ... }` points to `x`, but calls it an "array pattern."

Reviewed By: samwgoldman

Differential Revision: D23333225

fbshipit-source-id: 9896e3f739842dc74d440425dc0c7785a21bb4ee